refacto helm repository modal
