+++ libs/domains/service-helm/feature/src/lib/source-setting/source-setting.tsx
@@ -1,10 +1,12 @@
 import { Controller, useFormContext } from 'react-hook-form'
 import { useParams } from 'react-router-dom'
-import { InputSelect, InputText, LoaderSpinner } from '@qovery/shared/ui'
+import { HelmRepositoryCreateEditModal } from '@qovery/domains/organizations/feature'
+import { IconAwesomeEnum, IconFa, InputSelect, InputText, LoaderSpinner, useModal } from '@qovery/shared/ui'
 import { useHelmRepositories } from '../hooks/use-helm-repositories/use-helm-repositories'
 
 export function SourceSetting({ disabled = false }: { disabled?: boolean }) {
   const { organizationId = '' } = useParams()
+  const { openModal, closeModal } = useModal()
   const { control, watch } = useFormContext()
   const watchFieldProvider = watch('source_provider')
 
@@ -70,6 +72,18 @@ export function SourceSetting({ disabled = false }: { disabled?: boolean }) {
                     value={field.value}
                     error={error?.message}
                     isSearchable
+                    menuListButton={{
+                      title: 'Select helm repository',
+                      label: 'New helm repository',
+                      icon: <IconFa name={IconAwesomeEnum.CIRCLE_PLUS} className="text-brand-500" />,
+                      onClick: () => {
+                        openModal({
+                          content: (
+                            <HelmRepositoryCreateEditModal organizationId={organizationId} onClose={closeModal} />
+                          ),
+                        })
+                      },
+                    }}
                   />
                 )}
               />
